pyproject.toml
[project]
name = "fastapi-alembic-sqlmodel-poetry-starter"
version = "0.1.0"
description = "base poetry managed fastapi app"
authors = [
    {name = "Alhagie Bai Cham",email = "cham207388@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[tool.poetry.group.test.dependencies]
pytest = "^8.4.1"
six = "^1.17.0"
mockito = "^1.5.4"

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
